RequirementsCheck is not a task
